Remove conditional cluster name "onm3h-demo"
Removed the conditional check for the cluster name "onm3h-demo" from the widget  Node Memory Usage configuration. This condition caused the widget to be empty when the cluster name did not match "onm3h-demo". Each cluster has its own unique name, so this hard-coded condition was unnecessary and led to incorrect widget behavior.